Wordpress update, Timber integration

+++ wp-content/plugins/gravityforms/form_display.php
@@ -1,5 +1,9 @@
 <?php
 
+if(!class_exists('GFForms')){
+    die();
+}
+
 class GFFormDisplay{
 
     public static $submission = array();
@@ -9,6 +13,8 @@ class GFFormDisplay{
     const ON_CONDITIONAL_LOGIC = 2;
 
     public static function process_form($form_id){
+
+        GFCommon::log_debug( "Starting to process form (#{$form_id}) submission." );
 
         //reading form metadata
         $form = RGFormsModel::get_form_meta($form_id);
@@ -28,6 +34,8 @@ class GFFormDisplay{
         $page_number = $source_page_number;
         $target_page = self::get_target_page($form, $page_number, $field_values);
 
+        GFCommon::log_debug("Source page number: {$source_page_number}. Target page number: {$target_page}");
+
         //Loading files that have been uploaded to temp folder
         $files = GFCommon::json_decode(stripslashes(RGForms::post("gform_uploaded_files")));
         if(!is_array($files))
@@ -43,8 +51,13 @@ class GFFormDisplay{
             $is_valid = self::validate($form, $field_values, $page_number, $failed_validation_page);
         }
 
+        GFCommon::log_debug("After validation. Is submission valid? {$is_valid}");
+
         //Upload files to temp folder when going to the next page or when submitting the form and it failed validation
         if( $target_page >= $page_number || ($target_page == 0 && !$is_valid) ){
+
+            GFCommon::log_debug("Uploading files...");
+
             //Uploading files to temporary folder
             $files = self::upload_files($form, $files);
             RGFormsModel::$uploaded_files[$form_id] = $files;
@@ -61,6 +74,8 @@ class GFFormDisplay{
 
         $confirmation = "";
         if($is_valid && $page_number == 0){
+
+
             $ajax = isset($_POST["gform_ajax"]);
 
             //adds honeypot field if configured
@@ -70,11 +85,16 @@ class GFFormDisplay{
             $failed_honeypot = rgar($form,"enableHoneypot") && !self::validate_honeypot($form);
 
             if($failed_honeypot){
+
+                GFCommon::log_debug("Failed Honeypot validation. Displaying confirmation and aborting");
+
                 //display confirmation but doesn't process the form when honeypot fails
                 $confirmation = self::handle_confirmation($form, $lead, $ajax);
                 $is_valid = false;
             }
             else{
+
+                GFCommon::log_debug("Submission is valid. Moving forward.");
 
                 $form = self::update_confirmation($form);
 
@@ -173,15 +193,15 @@ class GFFormDisplay{
                         $price = GFCommon::to_number($price);
 
                         $product_name = !is_array($value) || empty($value[$field["id"] . ".1"]) ? $field["label"] : $value[$field["id"] . ".1"];
-                        $product_fields[$field["id"]. ".1"] = wp_hash($product_name);
-                        $product_fields[$field["id"]. ".2"] = wp_hash($price);
+                        $product_fields[$field["id"]. ".1"] = wp_hash(GFFormsModel::maybe_trim_input($product_name, $form["id"], $field));
+                        $product_fields[$field["id"]. ".2"] = wp_hash(GFFormsModel::maybe_trim_input($price, $form["id"], $field));
                     break;
 
                     case "singleshipping" :
                         $price = !empty($value) ? $value : $field["basePrice"];
                         $price = GFCommon::to_number($price);
 
-                        $product_fields[$field["id"]] = wp_hash($price);
+                        $product_fields[$field["id"]] = wp_hash(GFFormsModel::maybe_trim_input($price, $form["id"], $field));
                     break;
                     case "radio" :
                     case "select" :
@@ -191,7 +211,7 @@ class GFFormDisplay{
                             if($field["enablePrice"])
                                 $field_value .= "|" . GFCommon::to_number(rgar($choice,"price"));
 
-                            $product_fields[$field["id"]][] = wp_hash($field_value);
+                            $product_fields[$field["id"]][] = wp_hash(GFFormsModel::maybe_trim_input($field_value, $form["id"], $field));
                         }
                     break;
                     case "checkbox" :
@@ -204,16 +224,18 @@ class GFFormDisplay{
                             if($index % 10 == 0) //hack to skip numbers ending in 0. so that 5.1 doesn't conflict with 5.10
                                 $index++;
 
-                            $product_fields[$field["id"] . "." . $index++] = wp_hash($field_value);
+                            $product_fields[$field["id"] . "." . $index++] = wp_hash(GFFormsModel::maybe_trim_input($field_value, $form["id"], $field));
                         }
                     break;
 
                 }
             }
         }
-        $hash = serialize($product_fields);
+
+        $hash = json_encode($product_fields);
         $checksum = wp_hash(crc32($hash));
         return base64_encode(json_encode(array($hash, $checksum)));
+
     }
 
     private static function has_pages($form){
@@ -515,8 +537,8 @@ class GFFormDisplay{
                 else if($form["pagination"]["type"] == "steps"){
                     $form_string .="
                     <div id='gf_page_steps_{$form_id}' class='gf_page_steps'>";
-
-                    for($i=0, $count = sizeof($form["pagination"]["pages"]); $i<$count; $i++){
+                    $pages = isset($form["pagination"]["pages"]) ? $form["pagination"]["pages"] :  array();
+                    for($i=0, $count = sizeof($pages); $i<$count; $i++){
                         $step_number = $i+1;
                         $active_class = $step_number == $current_page ? " gf_step_active" : "";
                         $first_class = $i==0 ? " gf_step_first" : "";
@@ -530,7 +552,7 @@ class GFFormDisplay{
                         $classes = GFCommon::trim_all($classes);
 
                         $form_string .="
-                        <div id='gf_step_{$form_id}_{$step_number}' class='{$classes}'><span class='gf_step_number'>{$step_number}</span>&nbsp;{$form["pagination"]["pages"][$i]}</div>";
+                        <div id='gf_step_{$form_id}_{$step_number}' class='{$classes}'><span class='gf_step_number'>{$step_number}</span>&nbsp;{$pages[$i]}</div>";
                     }
 
                     $form_string .="
@@ -690,7 +712,7 @@ class GFFormDisplay{
                 //return regular confirmation message
                 if($ajax)
                 {
-                    $progress_confirmation = "<!DOCTYPE html><html><head><meta charset='UTF-8' /></head><body class='GF_AJAX_POSTBACK'>" . $confirmation_message . "</body></html>";
+                    $progress_confirmation = apply_filters('gform_ajax_iframe_content', "<!DOCTYPE html><html><head><meta charset='UTF-8' /></head><body class='GF_AJAX_POSTBACK'>" . $confirmation_message . "</body></html>");
                 }
                 else
                 {
@@ -709,6 +731,9 @@ class GFFormDisplay{
         $form_string = self::get_form_init_scripts($form);
         $current_page = self::get_current_page($form_id);
         $form_string .= "<script type='text/javascript'>" . apply_filters("gform_cdata_open", "") . " jQuery(document).ready(function(){jQuery(document).trigger('gform_post_render', [{$form_id}, {$current_page}]) } ); " . apply_filters("gform_cdata_close", "") . "</script>";
+
+        $form_string = apply_filters( 'gform_footer_init_scripts_filter', $form_string, $form, $current_page );
+        $form_string = apply_filters( 'gform_footer_init_scripts_filter_' . $form['id'], $form_string, $form, $current_page );
 
         if(!isset($_init_forms[$form_id])){
             echo $form_string;
@@ -827,6 +852,11 @@ class GFFormDisplay{
     }
 
     public static function is_empty($field, $form_id=0){
+
+        if(empty($_POST["is_submit_" . $field["formId"]])){
+            return true;
+        }
+
         switch(RGFormsModel::get_input_type($field)){
             case "post_image" :
             case "fileupload" :
@@ -893,15 +923,20 @@ class GFFormDisplay{
     }
 
 
-
+    /**
+     * Validates the range of the number according to the field settings.
+     *
+     * @param array $field
+     * @param array $value A decimal_dot formatted string
+     * @return true|false True on valid or false on invalid
+     */
     private static function validate_range($field, $value){
 
-        if( !GFCommon::is_numeric($value, rgar($field, "numberFormat")) )
+        if( !GFCommon::is_numeric($value, "decimal_dot") )
             return false;
 
-        $number = GFCommon::clean_number($value, rgar($field, "numberFormat"));
-        if( (is_numeric($field["rangeMin"]) && $number < $field["rangeMin"]) ||
-            (is_numeric($field["rangeMax"]) && $number > $field["rangeMax"])
+        if( (is_numeric($field["rangeMin"]) && $value < $field["rangeMin"]) ||
+            (is_numeric($field["rangeMax"]) && $value > $field["rangeMax"])
         )
             return false;
         else
@@ -939,6 +974,9 @@ class GFFormDisplay{
         //if Akismet plugin is installed, run lead through Akismet and mark it as Spam when appropriate
         $is_spam = GFCommon::akismet_enabled($form['id']) && GFCommon::is_akismet_spam($form, $lead);
 
+        GFCommon::log_debug("Checking for spam...");
+        GFCommon::log_debug("Is entry considered spam? {$is_spam}.");
+
         if(!$is_spam){
             GFCommon::create_post($form, $lead);
             //send notifications
@@ -962,18 +1000,25 @@ class GFFormDisplay{
             return false;
         $target_path = RGFormsModel::get_upload_path($form["id"]) . "/tmp/";
         $filename = $target_path . $unique_form_id . "_input_*";
-        array_map('unlink', glob($filename));
+        $files = glob($filename);
+        if (is_array($files)){
+        	array_map('unlink', $files);
+		}
 
         // clean up file from abandoned submissions older than 48 hours
         $files = glob($target_path."*");
-        foreach($files as $file) {
-            if(is_file($file) && time() - filemtime($file) >= 2*24*60*60) {
-                unlink($file);
-            }
-        }
+        if (is_array($files)){
+	        foreach($files as $file) {
+	            if(is_file($file) && time() - filemtime($file) >= 2*24*60*60) {
+	                unlink($file);
+	            }
+	        }
+		}
     }
 
     public static function handle_confirmation($form, $lead, $ajax=false){
+
+        GFCommon::log_debug("Sending confirmation");
 
 		//run the function to populate the legacy confirmation format to be safe
 		$form = self::update_confirmation($form, $lead);
@@ -1032,6 +1077,8 @@ class GFFormDisplay{
             $confirmation = self::get_js_redirect_confirmation($confirmation["redirect"], $ajax); //redirecting via client side
         }
 
+        GFCommon::log_debug("Confirmation: " . print_r($confirmation, true));
+
         return $confirmation;
     }
 
@@ -1068,6 +1115,10 @@ class GFFormDisplay{
         // validate entry limit
         if(self::validate_entry_limit($form))
             return false;
+
+        if(empty($_POST["is_submit_" . $form["id"]])){
+            return false;
+        }
 
         foreach($form["fields"] as &$field){
 
@@ -1112,7 +1163,7 @@ class GFFormDisplay{
                     break;
                 }
 
-                $field["validation_message"] =  apply_filters("gform_duplicate_message_{$form["id"]}", apply_filters("gform_duplicate_message", $default_message, $form), $form);
+                $field["validation_message"] = apply_filters( "gform_duplicate_message_{$form["id"]}", apply_filters("gform_duplicate_message", $default_message, $form, $field, $value ), $form, $field, $value );
 
             }
             else{
@@ -1226,6 +1277,8 @@ class GFFormDisplay{
                         break;
 
                         case "number" :
+
+                            // the POST value has already been converted from currency or decimal_comma to decimal_dot and then cleaned in get_field_value()
 
                             $value = GFCommon::maybe_add_leading_zero($value);
 
@@ -1529,11 +1582,11 @@ class GFFormDisplay{
             //making sure state wasn't tampered with by validating checksum
             $checksum = wp_hash(crc32($state[0]));
 
-            if($checksum != $state[1]){
+            if($checksum !== $state[1]){
                 return true;
             }
 
-            $_gf_state = unserialize($state[0]);
+            $_gf_state = json_decode($state[0], true);
         }
 
         if(!is_array($value)){
@@ -1589,10 +1642,17 @@ class GFFormDisplay{
 
     public static function enqueue_form_scripts( $form, $ajax = false ){
 
+        // adding pre enqueue scripts hook so that scripts can be added first if a need exists
+        do_action("gform_pre_enqueue_scripts_{$form["id"]}", do_action("gform_pre_enqueue_scripts", $form, $ajax), $ajax);
+
         if(!get_option('rg_gforms_disable_css')){
 
             wp_enqueue_style("gforms_reset_css", GFCommon::get_base_url() . "/css/formreset.css", null, GFCommon::$version);
-            wp_enqueue_style("gforms_datepicker_css", GFCommon::get_base_url() . "/css/datepicker.css", null, GFCommon::$version);
+
+            if( self::has_datepicker_field( $form ) ) {
+                wp_enqueue_style("gforms_datepicker_css", GFCommon::get_base_url() . "/css/datepicker.css", null, GFCommon::$version);
+            }
+
             wp_enqueue_style("gforms_formsmain_css", GFCommon::get_base_url() . "/css/formsmain.css", null, GFCommon::$version);
             wp_enqueue_style("gforms_ready_class_css", GFCommon::get_base_url() . "/css/readyclass.css", null, GFCommon::$version);
             wp_enqueue_style("gforms_browsers_css", GFCommon::get_base_url() . "/css/browsers.css", null, GFCommon::$version);
@@ -1611,7 +1671,7 @@ class GFFormDisplay{
             wp_enqueue_script( 'gform_datepicker_init' );
         }
 
-        if( $ajax || self::has_price_field( $form ) || self::has_password_strength( $form ) || GFCommon::has_list_field( $form ) || GFCommon::has_credit_card_field($form) || self::has_conditional_logic( $form ) || self::has_calculation_field( $form ) ) {
+        if( $ajax || self::has_price_field( $form ) || self::has_password_strength( $form ) || GFCommon::has_list_field( $form ) || GFCommon::has_credit_card_field($form) || self::has_conditional_logic( $form ) || self::has_currency_format_number_field($form) || self::has_calculation_field( $form ) ) {
             wp_enqueue_script( 'gform_gravityforms' );
         }
 
@@ -1655,13 +1715,17 @@ class GFFormDisplay{
             if( ! wp_style_is( 'gforms_css' ) ) {
 
                 wp_enqueue_style("gforms_reset_css", GFCommon::get_base_url() . "/css/formreset.css", null, GFCommon::$version);
-                wp_enqueue_style("gforms_datepicker_css", GFCommon::get_base_url() . "/css/datepicker.css", null, GFCommon::$version);
                 wp_enqueue_style("gforms_formsmain_css", GFCommon::get_base_url() . "/css/formsmain.css", null, GFCommon::$version);
                 wp_enqueue_style("gforms_ready_class_css", GFCommon::get_base_url() . "/css/readyclass.css", null, GFCommon::$version);
                 wp_enqueue_style("gforms_browsers_css", GFCommon::get_base_url() . "/css/browsers.css", null, GFCommon::$version);
 
                 wp_print_styles(array("gforms_reset_css"));
-                wp_print_styles(array("gforms_datepicker_css"));
+
+                if( self::has_datepicker_field( $form ) ) {
+                    wp_enqueue_style("gforms_datepicker_css", GFCommon::get_base_url() . "/css/datepicker.css", null, GFCommon::$version);
+                    wp_print_styles(array("gforms_datepicker_css"));
+                }
+
                 wp_print_styles(array("gforms_formsmain_css"));
                 wp_print_styles(array("gforms_ready_class_css"));
                 wp_print_styles(array("gforms_browsers_css"));
@@ -1710,8 +1774,9 @@ class GFFormDisplay{
 
         foreach( $scripts as $script ) {
             wp_enqueue_script( $script );
-            wp_print_scripts( array( $script ) );
         }
+
+        wp_print_scripts( $scripts );
 
         if( wp_script_is( 'gform_gravityforms' ) ) {
             echo '<script type="text/javascript"> ' . GFCommon::gf_global( false ) . ' </script>';
@@ -1719,7 +1784,7 @@ class GFFormDisplay{
 
     }
 
-    private static function has_conditional_logic( $form ) {
+    public static function has_conditional_logic( $form ) {
         $has_conditional_logic = self::has_conditional_logic_legwork( $form );
         return apply_filters( 'gform_has_conditional_logic', $has_conditional_logic, $form );
     }
@@ -2025,7 +2090,7 @@ class GFFormDisplay{
                 $currency_fields[] = "#input_{$form["id"]}_{$field["id"]}";
         }
 
-        return "gformInitCurrencyFormatFields('" . implode(",", $currency_fields) . "')";
+        return "gformInitCurrencyFormatFields('" . implode(",", $currency_fields) . "');";
     }
 
     public static function get_counter_init_script($form){
@@ -2106,7 +2171,7 @@ class GFFormDisplay{
                 continue;
 
             $mask = rgar($field, 'inputMaskValue');
-            $script = "jQuery('#input_{$form['id']}_{$field['id']}').mask('{$mask}').bind('keypress', function(e){if(e.which == 13){jQuery(this).blur();} } );";
+            $script = "jQuery('#input_{$form['id']}_{$field['id']}').mask('" . esc_js($mask). "').bind('keypress', function(e){if(e.which == 13){jQuery(this).blur();} } );";
 
             $script_str .= apply_filters("gform_input_mask_script_{$form['id']}", apply_filters("gform_input_mask_script", $script, $form['id'], $field['id'], $mask), $form['id'], $field['id'], $mask);
         }
@@ -2433,7 +2498,7 @@ class GFFormDisplay{
 
         $field_id = IS_ADMIN || $form_id == 0 ? "input_$id" : "input_" . $form_id . "_$id";
 
-        $required_div = IS_ADMIN || rgar($field, "isRequired") ? sprintf("<span class='gfield_required'>%s</span>", $field["isRequired"] ? "*" : "") : "";
+        $required_div = IS_ADMIN || rgar($field, "isRequired") ? sprintf("<span class='gfield_required'>%s</span>", rgar($field, "isRequired") ? "*" : "") : "";
         $target_input_id = "";
         $is_description_above = rgar($field, "descriptionPlacement") == "above";
